setup init package
