+++ README.md
@@ -2,3 +2,12 @@
 2. Go to directory "dist"
 3. Run .exe file "assistant"
 4. Follow instructions
+
+Hi, I'm a bot. And I can 
+<create>     save contacts with names, phone numbers, emails and birthdays to the contact book
+<birthday>   display a list of contacts who have a birthday after a specified number of days from the current date
+<search>     search for contacts and notes from the contact book
+<edit>       edit entries from the contact book, edit and delete notes
+<delete>     delete entries from the contact book
+<show>       display all contents of the contact book 
+<exit>       exit the program
