--- data/data-file.txt
@@ -1,4 +1,0 @@
-V| 5465654544| shg@adf.hg| 26.06.1999| 1111111111
-ss| 56464556456| jhghjf@hg.gf| 25.06.2015| aregvew 333
-asd| 65564645454| sdsg@gf.hf| 27.06.1990| jyuhfgjugh2222
-Aaa| 65465456465| nhgfh@jmhg.gf| 20.05.2005| show
