+++ personal_assistant/__init__.py
@@ -1,0 +1,1 @@
+from personal_assistant import assistant
